update_engine: Suspend update during restricted interval.
Make UpdateAttempter suspend download during restricted time
intervals from DeviceAutoUpdateTimeRestrictions policy. It may only
happen while DownloadAction is in progress.

When DownloadAction is started, UpdateAttempter creates
UpdateTimeRestrictionsMonitor which tracks the restricted time
intervals. The monitor informs the UpdateAttempter when a
restricted interval starts which then cancels the update process
if download is in progress. This monitor is destroyed when
DownloadAction is finished with or without error.

BUG=chromium:1117450
TEST=FEATURES=test emerge-${BOARD} update_engine

Change-Id: Ibd2a3291c251bc33618eb530f352f821ad5dcfd2
Reviewed-on: https://chromium-review.googlesource.com/c/aosp/platform/system/update_engine/+/2567121
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Tested-by: Saurabh Nijhara <snijhara@google.com>
Commit-Queue: Saurabh Nijhara <snijhara@google.com>