update_engine: Reorganize UpdateCheckAllowed policy
The current policy management is through class templates. It has been
quite daunting to deal with the way policies are worked around. Each
policy class has several functions and by addition of new functions, we
will be in a situation that each policy implementation will implemented
one or two functions and ignores the rest. Hence not a very good design.

In this CL we introduced a new PolicyEvaluator concept which is
repsonsible to take a policy and its associated data and evaulate that
policy. Policies can be recursive and one policy can evaluate other
policies too. The new design does not use class templates anymore hence
increases the readability and extensibility of the design.

BUG=b:179419726
TEST=cros_workon_make --board reef --test update_engine

Change-Id: If213fa2dac381ebc1ddb1eb7058aa45d6cb7a4c7
Reviewed-on: https://chromium-review.googlesource.com/c/aosp/platform/system/update_engine/+/1760445
Tested-by: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Commit-Queue: Amin Hassani <ahassani@chromium.org>