update_engine: Reroganize CanUpdateStart policy
This policy is actually not used anywhere. But there are good
information inside of it and it would be a good idea to merge its
concepts like scattering, etc into the UpdaetCanBeApplied policy.

Removed a check for UpdateCheckAllowed in this policy. In practice once
we hit the UpdateCanStartPolicy, we have already checked for an
update. So no need to do it again. This means some of the unittests have
to be deleted because they were not applicable anymore.

BUG=b:179419726
TEST=cros_workon_make --board reef --test update_engine

Change-Id: I23e73e20b07121ddb2eb1a1bdf3381c536fea8c3
Reviewed-on: https://chromium-review.googlesource.com/c/aosp/platform/system/update_engine/+/2681690
Commit-Queue: Amin Hassani <ahassani@chromium.org>
Tested-by: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Reviewed-by: Vyshu Khota <vyshu@chromium.org>