update_engine: Revert caching the update payload manifest
This CL basically revert this CL:
https://android-review.googlesource.com/c/platform/system/update_engine/+/1360759

The purpose of that CL was to cache the payload manifest in the
pref (wrong place) to read it next time the update is being
resumed. However, there seems to be some problems with it. For example
it is disregarding the p2p updates and is messing it up. The code seems
to be wrong (it weirdly creates a DeltaPerformer again). And it doesn't
have any unittests so we can't accept it in CrOS. This won't revert the
changes in the DownloadAction that Android uses, so it doesn't have any
effect on aosp.

BUG=b:171829801
TEST=cros_workon_make --board reef --test update_engine

Change-Id: Id1ed36a336970ffd119385dbe476f6c64103b407
Reviewed-on: https://chromium-review.googlesource.com/c/aosp/platform/system/update_engine/+/2776258
Tested-by: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Reviewed-by: Vyshu Khota <vyshu@google.com>
Commit-Queue: Amin Hassani <ahassani@chromium.org>