update_engine: Fix a bug with P2P updates when there are multiple payload
It seems currently, the implementaion opens the file descriptor for
writing the P2P file only once even if there are multiple payloads. This
is a bug for example where there are DLCs being updated in addition to
the platform update. This CL closes the FD everytime a payload is
sucessfully downloaded and opens it every time a new payload is starting
to be downloaded.

Unfortunately this can't be tested easily locally. So we need to land
this patch and then use DUTs in the lab to perform a p2p updates (using
canary builds) when DLCs are available to make sure this CL is correct.

BUG=b:172219304
TEST=cros_workon_make --board reef --test update_engine

Change-Id: If02035937288be84eb66a5c3abda87bce6ecdb09
Reviewed-on: https://chromium-review.googlesource.com/c/aosp/platform/system/update_engine/+/2776259
Tested-by: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Reviewed-by: Vyshu Khota <vyshu@chromium.org>
Commit-Queue: Amin Hassani <ahassani@chromium.org>