update_engine: Maintain a list of PolicyEvaluator
Currently, we don't keep any reference to a PolicyEvaluator simply
because it seemed like that's not needed as long as the message loop has
a task running that has some reference to the instance of
PolicyEvaluator which is corrects and works fine. However, the problem
is that the tasks that are scheduled by the PolicyEvaluator will get
desctructed really late in the shutdown. So the variables in the
Providers will get deleted before these tasks (and subsequently their
parent PolicyEvaluator) gets deleted. This causes some AsyncCopyVariable
type variables to have non-zero number of observers when getting deleted
and crash the update_engine on shutdown.

One simple solution might be to just forcefully remove all the observers
from variables that still have them. However, this might cause issues as
those observers might still be accessing these variables and could cause
further memory leak/access issues.

To fix this issue, we keep a list of active PolicyEvaluators in the
UpdateManager so they can get cleaned up on destruction of
UpdateManager. This also adds an unregister callback to PolicyEvaluators
so they can unregister themselves from UpdateManager when they go out of
scope.

BUG=b:183200527
TEST=cros_workon_make --board reef --test update_engine
TEST=cros deploy and restarted update-engine, it didn't crash anymore.

Change-Id: I9dd21ca67d18e54ca32807b774a9ba3fa81338e2
Reviewed-on: https://chromium-review.googlesource.com/c/aosp/platform/system/update_engine/+/2792467
Commit-Queue: Amin Hassani <ahassani@chromium.org>
Tested-by: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>