update_engine: Add OmahaRequestParamsPolicy
This policy is used to extract all the parameters needed for Omaha
requests. From this point forward if we need to send any new parameters
to the omaha that comes from device policy, it has to go through this
policy. We should eventually migrate all (or most) other values in
OmahaRequestParams through this class.

BUG=chromium:179419726
TEST=cros_workon_make --board reef --test update_engine

Change-Id: I8dbfc69a512ee165e78dfc45b01bb6c8c4338f9e
Reviewed-on: https://chromium-review.googlesource.com/c/aosp/platform/system/update_engine/+/2829355
Tested-by: Amin Hassani <ahassani@chromium.org>
Commit-Queue: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>