update_engine: Allow to disable sending market segment info
This CL allows disabling the market segment feature by accepting a
boolean flag from omaha response. The flag is named "_disable_dms" and
should be an attribute of <updatecheck> tag. If this flag is set to
true, a new boolean pref "market-segment-disabled" is written and
consequent update checks will omit adding market segment information in
the update check and ping requests until it is not disabled anymore.

BUG=chromium:1192372
TEST=cros_workon_make --board reef --test update_engine
TEST=cros deploy and did update_engine_client --check_for_update; the market_semgnet="consumer" was in the <os> tag.
TEST=wrote "true" to file
/var/lib/update_engine/prefs/market-segment-disabled, restarted update-engine, and did update check, no market_segment tag was being sent.

Change-Id: I9ea3ce74f429d2790e1323057aa687d294071be8
Reviewed-on: https://chromium-review.googlesource.com/c/aosp/platform/system/update_engine/+/2829356
Tested-by: Amin Hassani <ahassani@chromium.org>
Commit-Queue: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>