update_engine: Disable p2p when the device is has no owner
Or when it has owner but the owner is empty (enterprise
enrolled). Historically, it seems like it has always been like
this (introduced in crrev.com/c/221476). But it doesn't really make
sense. If the policy tells no p2p, we should not enable based on the
device ownership result. Furthermore, this is causing the sharing of p2p
be ON, but downloading it be OFF. Means, for example in Google, everyone
is sharing p2p files, but no body is allowed to download it. So it's
better to enable p2p downloads purely by policy or if explicity
requested by DBus (mostly for testing).

BUG=b:186147823
TEST=cros_workon_make --board zork --test update_engine

Change-Id: Ib3544b8282aa9a28c6561efeeaa5327cd9701a8c
Reviewed-on: https://chromium-review.googlesource.com/c/aosp/platform/system/update_engine/+/2848117
Tested-by: Amin Hassani <ahassani@chromium.org>
Commit-Queue: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>