update_engine: Reset current slot to be the active slot when downloading
This makes sure that the current slot is always the active slot, since
there are times where this might not be the case, such as when
consecutive updates complete and change the active slot. Setting it
back to the current slot when a download starts make sure the system
is always booting to a valid image.

BUG=b:184899345
TEST=FEATURES=test update_engine, enable repeated updates on a DUT
restart in the middle of the second update.

Change-Id: I9c1f90ca22913aa0358e10b180a54774f6f465af
Reviewed-on: https://chromium-review.googlesource.com/c/aosp/platform/system/update_engine/+/2882498
Tested-by: Vyshu Khota <vyshu@google.com>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Reviewed-by: Amin Hassani <ahassani@chromium.org>
Commit-Queue: Amin Hassani <ahassani@chromium.org>