update_engine: Do not reset to IDLE if waiting for reboot.
If a consective update check returns kNoUpdate do not reset the
current status to IDLE. A previous update still needs to be applied
by rebooting. Look at the update complete marker to determine what to
reset the status to. The marker is erased everytime a new update starts
and is written after an update is completed. If is exists, there is
a update waiting to be applied.

BUG=b:187831741
TEST=FEATURES=test emerge-zork update_engine
TEST= cros deploy update_engine, test repeated updates.
restart update-engine


Change-Id: Id40f977cc95d04070037406fec37fdb92d716526
Reviewed-on: https://chromium-review.googlesource.com/c/aosp/platform/system/update_engine/+/2896249
Tested-by: Vyshu Khota <vyshu@chromium.org>
Commit-Queue: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>