update engine: Read and parse kernel configs.
Use `dump_kernel_config` to extract the kernel flags for MiniOS
partitions.
This function searches through the configs for a given flag and
returns the value. In later CLs this is used to get the version
to send with the update check.

BUG=b:186486897
TEST=cros deploy ssh://<IP> update_engine.
Verify the flag is parsed.

Disallow-Recycled-Builds: all
Change-Id: I96f85fae6e2ba5e4ff350dce8a3c963e613645f7
Reviewed-on: https://chromium-review.googlesource.com/c/aosp/platform/system/update_engine/+/2997168
Commit-Queue: Vyshu Khota <vyshu@chromium.org>
Tested-by: Vyshu Khota <vyshu@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>