update_engine: Add target_version_selector to Omaha request.
Add target_version_selector policy to OmahaRequestParams. The
corresponding request field is targetversionselector.

The policy is passed as is without being processed on the device and
should only be interpreted by Omaha. So the policy value and format are
free for change.

BUG=chromium:1237021
TEST=cros_workon_make --board $BOARD --test chromeos-base/update_engine

Cq-Depend: chromium:3187090
Change-Id: If49a5b7b7c06fea0946c29e5f524000c4fcb3f4e
Reviewed-on: https://chromium-review.googlesource.com/c/aosp/platform/system/update_engine/+/3187093
Tested-by: Artem Sumaneev <asumaneev@google.com>
Tested-by: Victor-Gabriel Savu <vsavu@google.com>
Commit-Queue: Victor-Gabriel Savu <vsavu@google.com>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>