update_engine: Ignore update on metered network
Read the metered property from shill to detect if the current network is
metered. Changes state to NEED_PERMISSION_TO_UPDATE and ignores the
update attempt if it is metered. Otherwise, continue to check network
type, device policy and preference to see if update is allowed.

BUG=b:173734506
TEST=cros_workon_make --board=$B update_engine --test

Change-Id: I0728deff252b76388bc5002fef30de1a6907f08d
Reviewed-on: https://chromium-review.googlesource.com/c/aosp/platform/system/update_engine/+/3271435
Tested-by: Yuanpeng Ni <yuanpengni@google.com>
Commit-Queue: Yuanpeng Ni <yuanpengni@google.com>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Reviewed-by: Henry Barnor <hbarnor@chromium.org>