update_engine: Add toggle for consumer auto update
This is a new policy for checking consumer auto update feature.
The reason for checking this as an async variable is to allow for
changes to the feature to cause another policy evaluation.

The features that can be toggled are also cleaned up in how they are
managed.
Note: Future toggles and checks should be written as variables to be
saftely used within policies owned by providers.

BUG=chromium:1278079
TEST=FEATURES=test emerge-$B update_engine system_api
TEST=# test interval for periodic check.
TEST=# disable/enable consumer auto update feature.

Cq-Depend: chromium:3400218
Change-Id: Iab7d3e25d4cb7d802ba0d4b10b8bb0d1d6a64d01
Reviewed-on: https://chromium-review.googlesource.com/c/aosp/platform/system/update_engine/+/3390095
Tested-by: Jae Hoon Kim <kimjae@chromium.org>
Reviewed-by: Henry Barnor <hbarnor@chromium.org>
Commit-Queue: Jae Hoon Kim <kimjae@chromium.org>