update_engine: Add will_defer_update to update engine status
Whether an update will be deferred or not is determined at the beginning
of the update, and it won't be changed by toggling the deferred update
feature during the process. Therefore, include it in the update engine
status, so the UI gets this information from update engine instead of
relying on the toggle status.

BUG=b:244480509
TEST=FEATURES=test emerge-$B update_engine

Cq-Depend: chromium:3936452
Change-Id: Ida3fb7aeddd22f21e34b9f19c48fd3fd76bce1f3
Reviewed-on: https://chromium-review.googlesource.com/c/aosp/platform/system/update_engine/+/3934525
Reviewed-by: Henry Barnor <hbarnor@chromium.org>
Tested-by: Yuanpeng Ni‎ <yuanpengni@chromium.org>
Commit-Queue: Yuanpeng Ni‎ <yuanpengni@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>