update_engine: Don't update when resuming from hibernate
When resuming from hibernate, we are about to jump into a "different"
kernel, with a different mount namespace, and different instance of apps
running. Applying a system update during this resume would be a bad
idea, as the resumed system we may eventually jump into would have no
knowledge of this update (or partial update).

Add a policy that blocks updates when the system is resuming from
hibernate. A future change will make this time limited so that updates
cannot be deferred indefinitely by sitting on the resumed login screen.

BUG=b:237090293
TEST=Manipulate /run/hibernate/resume_in_progress, see updates defer

Change-Id: Idd79099514ef7772ae7906554c4394eaafa4003f
Reviewed-on: https://chromium-review.googlesource.com/c/aosp/platform/system/update_engine/+/3964539
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Tested-by: Evan Green <evgreen@chromium.org>
Commit-Queue: Evan Green <evgreen@chromium.org>