update_engine: Time limit hibernate resume hold policy
Resuming from hibernation holds off updates so as not to confuse the
resumed system. However, we don't want the resume to be able to hold
updates off indefinitely. Add a time limit so that updates are still
applied in a timely manner. If the time limit expires and a resume is
still pending, we abort the resume from hibernate and allow the update
to apply.

Resuming from hibernate is a state systems only ever boot into, they
don't transition from not-resuming to resuming within a boot. So the
timeout duration could be a measurement of boot time. Use the updater
uptime as a reasonable proxy for boot time. If the updater were to
restart, the timer would reset, but that shouldn't be a huge deal.

The HibernateInterface now potentially needs to fire off a dbus call to
abort the hibernate. Add a USE flag for hibernate, and a stub
implementation if the hiberman-client is unavailable. We also abuse the
var_ interface a bit, as fetching the abort_resume_from_hibernate var
is solely a side effect producer that kicks off the abort. There seems
to be some precedent for this, as the random_provider's var is
essentially a sidechannel to call getrandom().

BUG=b:237090293
TEST=Compile with low limit, observe abort gets sent.

Cq-Depend: chromium:3964511, chromium:3963981
Change-Id: I292d2c9ba5ebf7abf0a81131ad37cab945acb5c0
Reviewed-on: https://chromium-review.googlesource.com/c/aosp/platform/system/update_engine/+/3964540
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Commit-Queue: Evan Green <evgreen@chromium.org>
Tested-by: Evan Green <evgreen@chromium.org>